+++ stories/patterns/filters/filter-include-exclude/docs/filter-include-exclude.stories.tsx
@@ -1,0 +1,13 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import FiltersIncludeExcludeExample from './examples/basic-example';
+
+const meta: Meta = {
+  title: 'Patterns/Filters/FiltersIncludeExclude',
+};
+
+export default meta;
+
+export const FiltersIncludeExclude: StoryObj = {
+  render: FiltersIncludeExcludeExample,
+};