+++ semcore/notice/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.43.0] - 2024-11-29
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/animation` [2.38.2 ~> 2.39.0], `@semcore/utils` [4.43.3 ~> 4.44.0], `@semcore/button` [5.39.4 ~> 5.40.0], `@semcore/flex-box` [5.38.2 ~> 5.39.0], `@semcore/icon` [4.52.2 ~> 4.53.0], `@semcore/core` [2.36.2 ~> 2.37.0]).
+
 ## [5.42.4] - 2024-11-22
 
 ### Changed